chore: shift stuff out for reusability
